Atualização de Atributos nas classe.
