--- src/main/java/com/Reserva/Restaurante/domain/adm/DadosUpdateAdm.java
@@ -1,9 +1,0 @@
-package com.Reserva.Restaurante.domain.adm;
-
-public record DadosUpdateAdm(
-        String nome,
-        String telefone,
-        String email,
-        String senha
-) {
-}
