+++ src/main/java/com/Reserva/Restaurante/domain/cliente/DadosCadastroCliente.java
@@ -1,0 +1,6 @@
+package com.Reserva.Restaurante.domain.cliente;
+
+public record DadosCadastroCliente(
+
+) {
+}
