--- src/main/java/com/Reserva/Restaurante/domain/clientes/DadosCadastroCliente.java
@@ -1,6 +1,0 @@
-package com.Reserva.Restaurante.domain.clientes;
-
-public record DadosCadastroCliente(
-
-) {
-}
