--- src/main/java/com/Reserva/Restaurante/domain/usuarios/DadosCadastroUsuario.java
@@ -1,10 +1,0 @@
-package com.Reserva.Restaurante.domain.usuarios;
-
-import java.time.LocalDate;
-
-public record DadosCadastroUsuario(
-        String userName,
-        String senha,
-        LocalDate dataCriacao
-) {
-}
