--- src/main/java/com/Reserva/Restaurante/domain/usuarios/DadosListagemUsuario.java
@@ -1,14 +1,0 @@
-package com.Reserva.Restaurante.domain.usuarios;
-
-import java.time.LocalDate;
-
-public record DadosListagemUsuario(
-        Long id,
-        String userName,
-        String senha,
-        LocalDate dataCriacao
-) {
-    public DadosListagemUsuario(Usuario usuario){
-        this(usuario.getId(), usuario.getUserName(),usuario.getSenha(),usuario.getDataCriacao());
-    }
-}
