--- src/main/java/com/Reserva/Restaurante/domain/usuarios/DadosUpdateUsuario.java
@@ -1,10 +1,0 @@
-package com.Reserva.Restaurante.domain.usuarios;
-
-import java.time.LocalDate;
-
-public record DadosUpdateUsuario(
-        String userName,
-        String senha,
-        LocalDate dataCriacao
-) {
-}
