--- src/main/java/com/Reserva/Restaurante/repository/AdmRepository.java
@@ -1,7 +1,0 @@
-package com.Reserva.Restaurante.repository;
-
-import com.Reserva.Restaurante.domain.adm.Adm;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AdmRepository extends JpaRepository<Adm, Long> {
-}
