+++ src/main/java/com/Reserva/Restaurante/repository/ClientesRepository.java
@@ -1,7 +1,7 @@
 package com.Reserva.Restaurante.repository;
 
-import com.Reserva.Restaurante.domain.clientes.Clientes;
+import com.Reserva.Restaurante.domain.cliente.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientesRepository extends JpaRepository<Clientes, Long> {
+public interface ClientesRepository extends JpaRepository<Cliente, Long> {
 }
