+++ src/main/java/com/Reserva/Restaurante/repository/ReservaRepository.java
@@ -1,8 +1,9 @@
 package com.Reserva.Restaurante.repository;
 
 
-import com.Reserva.Restaurante.domain.reservas.Reservas;
+import com.Reserva.Restaurante.domain.reserva.Reserva;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReservaRepository extends JpaRepository<Reservas, Long> {
+public interface ReservaRepository extends JpaRepository<Reserva, Long> {
 }
