--- src/main/java/com/Reserva/Restaurante/repository/UsuariosRepository.java
@@ -1,8 +1,0 @@
-package com.Reserva.Restaurante.repository;
-
-import com.Reserva.Restaurante.domain.usuarios.Usuario;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UsuariosRepository extends JpaRepository<Usuario, Long> {
-
-}
