+++ src/main/java/com/Reserva/Restaurante/service/ClienteService.java
@@ -1,0 +1,14 @@
+package com.Reserva.Restaurante.service;
+
+import com.Reserva.Restaurante.repository.ClientesRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ClienteService {
+    @Autowired
+    private ClientesRepository clientesRepository;
+
+    //criar Clientes
+
+}
