--- src/main/java/com/Reserva/Restaurante/service/ClientesService.java
@@ -1,14 +1,0 @@
-package com.Reserva.Restaurante.service;
-
-import com.Reserva.Restaurante.repository.ClientesRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class ClientesService {
-    @Autowired
-    private ClientesRepository clientesRepository;
-
-    //criar Clientes
-
-}
